# docs/openapi.yml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    REST API for the Stellar GreenPay climate donation platform.
    Every donation is recorded on the Stellar blockchain via Soroban smart contracts.
    All `/api/v1/*` JSON responses use a single envelope. Successful responses
    are `{ success: true, data, meta? }`; error responses are
    `{ success: false, error: { code, message, details? } }`. The unversioned
    `/api/*` paths are legacy aliases that redirect to `/api/v1/*`; clients
    should migrate to `/api/v1/*` and branch on `error.code`, never message text.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  schemas:
    Success:
      type: object
      required: [success, data]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [true]
          example: true
        data:
          description: Endpoint-specific response payload.
        meta:
          type: object
          additionalProperties: true
          description: Optional response metadata such as pagination or cursors.
    Error:
      type: object
      required: [success, error]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [false]
          example: false
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              example: PROJECT_NOT_FOUND
            message:
              type: string
              description: Human-readable display message.
              example: Project not found
            details:
              description: Optional structured diagnostic details.
...
    ProjectUpdate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/updates/{projectId}:
    get:
      tags: [Updates]
      summary: List updates for a project
      operationId: listProjectUpdates
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated project updates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProjectUpdate'
                      nextCursor:
                        type: string
                        format: date-time
                        nullable: true
  /api/updates:
    post:
      tags: [Updates]
      summary: Post a project update (admin)
      operationId: createProjectUpdate
      parameters:
        - name: x-admin-key
          in: header
          description: Admin API key (required when ADMIN_API_KEY env var is set)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, title, body]
              properties:
                projectId:
                  type: string
                  format: uuid
                title:
                  type: string
                body:
                  type: string
      responses:
        '201':
          description: Update created and subscribers notified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectUpdate'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/updates/{updateId}/like:
    post:
      tags: [Updates]
      summary: Toggle like on an update
      operationId: toggleUpdateLike
      parameters:
        - name: updateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [donorAddress]
              properties:
                donorAddress:
                  type: string
      responses:
        '200':
          description: Like toggled
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          liked:
                            type: boolean
                          likeCount:
                            type: integer
        '400':
          description: Missing donorAddress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Update not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/updates/{updateId}/likes:
    get:
      tags: [Updates]
      summary: Get like count and user like status for an update
      operationId: getUpdateLikes
      parameters:
        - name: updateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: donorAddress
          in: query
          description: Check if this address has liked the update
          schema:
            type: string
      responses:
        '200':
          description: Like info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          likeCount:
                            type: integer
                          liked:
                            type: boolean

  # ── Subscriptions ────────────────────────────────────────────────────────────
...